Insets for the card reader activity to make it look ok on the gesture navigation
