+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/root/WooPosActivity.kt
@@ -1,9 +1,7 @@
 package com.woocommerce.android.ui.woopos.root
 
-import android.content.Context
 import android.content.pm.ActivityInfo
 import android.os.Bundle
-import android.util.TypedValue
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.foundation.layout.navigationBarsPadding
@@ -18,6 +16,7 @@ import androidx.core.view.WindowInsetsCompat
 import com.woocommerce.android.ui.woopos.cardreader.WooPosCardReaderFacade
 import com.woocommerce.android.ui.woopos.common.composeui.WooPosTheme
 import com.woocommerce.android.ui.woopos.support.WooPosGetSupportFacade
+import com.woocommerce.android.ui.woopos.util.ext.isGestureNavigation
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
@@ -68,18 +67,4 @@ private fun Modifier.gesturesOrButtonsNavigationPadding(): Modifier {
     } else {
         this.navigationBarsPadding()
     }
-}
-
-// That seems to be different on different devices, but 32dp is a common upper value
-private const val GESTURE_NAVIGATION_BAR_HEIGHT_DP = 32
-private fun WindowInsetsCompat.isGestureNavigation(context: Context): Boolean {
-    val bottomInset = getInsets(WindowInsetsCompat.Type.navigationBars()).bottom
-
-    val gestureNavigationBarHeightPx = TypedValue.applyDimension(
-        TypedValue.COMPLEX_UNIT_DIP,
-        GESTURE_NAVIGATION_BAR_HEIGHT_DP.toFloat(),
-        context.resources.displayMetrics
-    ).toInt()
-
-    return bottomInset in 1..gestureNavigationBarHeightPx
 }
